+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # ts2workflows changelog
 
+## <unreleased>
+
+Fixes:
+
+- Type annotation fixes in workflowslib
+
 ## Version 0.14.0 - 2025-11-19
 
 Fixes: